+++ invaders v2/Screen.h
@@ -11,10 +11,10 @@ protected:
 public:
 	Screen(e::XMVECTOR pos);
 	void Render(const RenderParams &params);
-	int Loop(int input, float frame);
+	int Loop(int input, int frame);
 	Camera &GetCamera(){ return camera; }
 protected:
 	virtual void RenderInternal(const RenderParams &params) = 0;
-	virtual int LoopInternal(int input, float frame) = 0;
+	virtual int LoopInternal(int input, int frame) = 0;
 	virtual void DelayInternal() = 0;
 };
